Refactor storage to mirror `.py` structure
