+++ django-stubs/core/files/storage/handler.pyi
@@ -1,0 +1,14 @@
+from typing import Any
+
+from django.core.exceptions import ImproperlyConfigured
+
+from .base import Storage
+
+class InvalidStorageError(ImproperlyConfigured): ...
+
+class StorageHandler:
+    def __init__(self, backends: dict[str, Storage] | None = ...) -> None: ...
+    @property
+    def backends(self) -> dict[str, Storage]: ...
+    def __getitem__(self, alias: str) -> Storage: ...
+    def create_storage(self, params: dict[str, Any]) -> Storage: ...
